requirements.txt
Flask==2.3.2
bcrypt==4.0.1
Werkzeug==2.3.6
PyJWT==2.8.0
bleach==6.0.0
